packages/rig-cli/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rig-cli"
version = "0.1.0"
description = "RIG CLI"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Apache-2.0"}
authors = [{name = "RIG Contributors"}]
dependencies = [
  "typer>=0.12.0",
  "rich>=13.7.0",
  "httpx>=0.27.0",
  "PyYAML>=6.0.1",
  "uvicorn>=0.23.0",
  "rig-core>=0.1.0",
  "rig-protocol-rgp>=0.1.0"
]

[project.scripts]
rig = "rig_cli.main:app"

[tool.setuptools]
packages = ["rig_cli"]
